feat: adjust needs explorer layout, fix needs overview data issues, update region / subregion needs links

+++ gatsby/customize-schema.ts
@@ -134,7 +134,7 @@ export default {
         survey: DANeedSurvey!
         place: DANeedPlace!
         product: DAProduct!
-        need: Float
+        need: Int
       }
     `
